requirements.txt
# Web框架 - 用于构建网站后端
Flask==3.0.0

# HTTP请求库 - 用于调用飞书API
requests==2.31.0

# 环境变量管理 - 用于从.env文件加载配置
python-dotenv==1.0.0

# Flask缓存扩展 - 用于缓存API结果，提高性能
Flask-Caching==2.0.2

# Flask依赖项 - 以下是Flask框架所需的依赖包
Werkzeug==3.0.1  # WSGI工具库
Jinja2==3.1.2    # 模板引擎
MarkupSafe==2.1.3  # 安全HTML字符串处理
itsdangerous==2.1.2  # 数据签名工具
click==8.1.7  # 命令行接口工具
blinker==1.6.2  # 信号处理库

# 时区处理库 - 用于处理日期和时间
pytz==2023.3